spotlights with helpers added
